refactor: clean and optimize code

+++ api/useCases/DeleteUser/DeleteUserUseCase.ts
@@ -4,6 +4,10 @@ export class DeleteUserUseCase {
   constructor(private userRepository: IUserRepository) {}
 
   async execute(id: string) {
-    return this.userRepository.deleteUser(id)
+    const userExists: boolean = await this.userRepository.verifyIdUserExists(id)
+    if(!userExists){
+      throw new Error('Esse usuário não existe!');
+    }
+    return await this.userRepository.deleteUser(id)
   }
 }
